--- tests/Browser/Form/InputComponents/InputComponent.php
@@ -1,28 +1,0 @@
-<?php
-
-namespace Tests\Browser\Form\InputComponents;
-
-use Livewire\Component;
-
-class InputComponent extends Component
-{
-    public ?string $entangle = null;
-
-    public function render(): string
-    {
-        return <<<'HTML'
-        <div>
-            <p dusk="entangled">{{ $entangle }}</p>
-       
-            <x-input dusk="entangle" label="Foo" hint="Bar" wire:model="entangle" />
-            
-            <x-button dusk="sync-entangle" wire:click="sync">Sync</x-button>
-        </div>
-HTML;
-    }
-
-    public function sync(): void
-    {
-        //
-    }
-}
