+++ tests/Browser/Form/RadioTest.php
@@ -2,38 +2,74 @@
 
 namespace Tests\Browser\Form;
 
-use Laravel\Dusk\Browser;
+use Livewire\Component;
+use Livewire\Livewire;
 use Tests\Browser\BrowserTestCase;
-use Tests\Browser\Form\RadioComponents\RadioComponent;
-use Tests\Browser\Form\RadioComponents\RadioLiveEntangleComponent;
 
 class RadioTest extends BrowserTestCase
 {
     /** @test */
     public function can_entangle(): void
     {
-        $this->browse(function (Browser $browser) {
-            $this->visit($browser, RadioComponent::class)
-                ->assertSee('Receive Alert')
-                ->click('@entangle-true')
-                ->click('@sync-entangle')
-                ->waitForTextIn('@entangled', 'true')
-                ->click('@entangle-false')
-                ->click('@sync-entangle')
-                ->waitForTextIn('@entangled', 'false');
-        });
+        Livewire::visit(new class extends Component
+        {
+            public ?bool $entangle = null;
+
+            public function render(): string
+            {
+                return <<<'HTML'
+                <div>
+                    <p dusk="entangled">@json($entangle)</p>
+
+                    <form>
+                        <x-radio dusk="entangle-true" wire:model="entangle" label="Receive Alert" value="1" />
+                        <x-radio dusk="entangle-false" wire:model="entangle" label="Receive Alert" value="0" />
+                    </form>
+                    
+                    <x-button dusk="sync-entangle" wire:click="sync">Sync</x-button>
+                </div>
+                HTML;
+            }
+
+            public function sync(): void
+            {
+                //
+            }
+        })
+            ->assertSee('Receive Alert')
+            ->click('@entangle-true')
+            ->click('@sync-entangle')
+            ->waitForTextIn('@entangled', 'true')
+            ->click('@entangle-false')
+            ->click('@sync-entangle')
+            ->waitForTextIn('@entangled', 'false');
     }
 
     /** @test */
     public function can_live_entangle(): void
     {
-        $this->browse(function (Browser $browser) {
-            $this->visit($browser, RadioLiveEntangleComponent::class)
-                ->assertSee('Receive Alert')
-                ->click('@entangle-true')
-                ->waitForTextIn('@entangled', 'true')
-                ->click('@entangle-false')
-                ->waitForTextIn('@entangled', 'false');
-        });
+        Livewire::visit(new class extends Component
+        {
+            public ?bool $entangle = null;
+
+            public function render(): string
+            {
+                return <<<'HTML'
+                <div>
+                    <p dusk="entangled">@json($entangle)</p>
+
+                    <form>
+                        <x-radio dusk="entangle-true" wire:model.live="entangle" label="Receive Alert" value="1" />
+                        <x-radio dusk="entangle-false" wire:model.live="entangle" label="Receive Alert" value="0" />
+                    </form>
+                </div>
+                HTML;
+            }
+        })
+            ->assertSee('Receive Alert')
+            ->click('@entangle-true')
+            ->waitForTextIn('@entangled', 'true')
+            ->click('@entangle-false')
+            ->waitForTextIn('@entangled', 'false');
     }
 }
